# perf: use number constructor